+++ src/prognostic_equations/remaining_tendency.jl
@@ -336,6 +336,14 @@ NVTX.@annotate function additional_tendency!(Yₜ, Y, p, t)
     chd = p.atmos.constant_horizontal_diffusion
     horizontal_constant_diffusion_tendency!(Yₜ, Y, p, t, chd)
 
+    edmfx_sgs_horizontal_diffusive_flux_tendency!(
+        Yₜ,
+        Y,
+        p,
+        t,
+        p.atmos.turbconv_model,
+    )
+
     # Optional tendency to bring negative small tracers back from negative
     # at the cost of water vapor.
     tracer_nonnegativity_vapor_tendency!(Yₜ, Y, p, t, microphysics_model)